Create Challenge #3
This function takes two parameters ,the object and the key. The function first checks if the key exists in the object. If it does, the function gets the value of the key. If the value is an object, the function reversely calls itself to get the value of the key from the nested object. Otherwise, the function returns the value of the key.